Change the way error messages are displayed
This commit will allow the program to display custom error messages
based on the type of error found.
When an error is detected, an event is triggered. The ErrorDisplayView
will listen to the event call and display an error message.
The error message will depend on the ErrorCode passed as an argument
when an error is detected.